==> W1_ML_on_simulated_data/1.Preprocess.ipynb <==
# import numpy
import pandas as pd

file_path = './simulated_survival_data'

'''gene expression: 20 genes are all features'''
gene_exp_train = pd.read_csv(f'{file_path}/X_train.csv', header=None)
gene_exp_test = pd.read_csv(f'{file_path}/X_test.csv', header=None)

'''survive:1=die/censor, 0=live; time(unit: months)'''
survive_train = pd.read_csv(f'{file_path}/Y_train.csv')
survive_test = pd.read_csv(f'{file_path}/Y_test.csv')

'''treatment: 1=Drug use, 0=Placebo use'''
treatment_train = pd.read_csv(f'{file_path}/TREATMENT_train.csv',header=None)
treatment_test = pd.read_csv(f'{file_path}/TREATMENT_test.csv',header=None)

treatment_train

def preprocess_steps(gene_exp_df,survive_df,treatment_df):
    '''
    1.concat three dataset
    2.screening control arm groups
    '''
    data_df = pd.concat([gene_exp_df, survive_df, treatment_df], axis=1)
    columns = list(data_df.columns)
    columns[-1] = 'treatment'
    data_df.columns = columns
    # print(data_df.columns)
    
    data_df = data_df[data_df['treatment']==0]
    data_df = data_df.reset_index(inplace=False, drop=True)
    print(data_df.head)
    return data_df

train_df = preprocess_steps(gene_exp_train, survive_train, treatment_train)
test_df = preprocess_steps(gene_exp_test, survive_test, treatment_test)

train_df

test_df

train_df.to_csv(f'{file_path}/control_train.csv')
test_df.to_csv(f'{file_path}/control_test.csv')

train_df['time'].describe()

train_df['event'].describe()

print((train_df['event']==0.0).sum())
print((train_df['event']==1.0).sum())

print((test_df['event']==0.0).sum())
print((test_df['event']==1.0).sum())
