==> W1_ML_on_simulated_data/2.Model Building.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import KFold
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier

file_path = './simulated_survival_data'
train_df_control = pd.read_csv(f'{file_path}/control_train.csv', index_col=0)
X_train = train_df_control.iloc[:, :20] #gene exp
Y_train = train_df_control.iloc[:, 20:22] #survival time and event

print(X_train.shape, Y_train.shape)

# clean data

invalid_count = (Y_train['time'] <= 0).sum()
valid_mask = Y_train['time'] > 0

X_filtered = X_train.loc[valid_mask].copy()
y_filtered = Y_train.loc[valid_mask].copy()

X = X_filtered
y = y_filtered

print(X.shape, y.shape)

## 1. prepare scikit-survival format

from sksurv.util import Surv

y_structured = Surv.from_arrays(
    event=y_filtered['event'].values.astype(bool),
    time=y_filtered['time'].values
)
# y_structured = Surv.from_arrays(event=y['event'].values.astype(bool), time=y['time'].values)
lifelines_df = pd.concat([X, y], axis=1)

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
X_scaled = pd.DataFrame(scaler.fit_transform(X), columns=X.columns)

print(X_scaled.head)

X_scaled.to_csv(f'{file_path}/X_scaled.csv')


#### 1. Random Forest (RF)
RF is often used for classification or regression, where we use it as a classification model to predict whether an event will occur (event).

`n_estimators`: The number of trees

`random_state`: to ensure that the result can be reproduced


from sksurv.ensemble import RandomSurvivalForest

rsf_model = RandomSurvivalForest(n_estimators=100, random_state=42)
rsf_model.fit(X_scaled, y_structured)

risk_scores_rsf = rsf_model.predict(X_scaled)

print(risk_scores_rsf[:5])

#### 2. Cox Regression (CoxPH)

Cox Proportional risk model, which is the standard survival analysis model

Note: The model of the lifelines library usually requires a single DataFrame that contains all the data (features, time, and events).

instead of separate X and y, and it is insensitive to data scaling

from lifelines import CoxPHFitter

cph = CoxPHFitter()
cph.fit(lifelines_df, duration_col='time', event_col='event')

cph.print_summary()

#### 3.Aalen's Additive Model

Unlike the Cox model, it assumes that the effects of covariates are additive rather than multipliative.

from lifelines import AalenAdditiveFitter

aaf = AalenAdditiveFitter(fit_intercept=False)
aaf.fit(lifelines_df, duration_col='time', event_col='event')

aaf.print_summary()

#### 4.Survival Support Vector Model

Apply the principles of support vector machines (SVMs) to survival data. It aims to find a hyperplane that maximizes the difference between the time of occurrence of an event and the time of review.

from sksurv.svm import FastSurvivalSVM

svm_model = FastSurvivalSVM(random_state=42)
svm_model.fit(X_scaled, y_structured)
risk_scores_svm = svm_model.predict(X_scaled)

risk_scores_svm[:5]

#### 5.XGBoost for survival version

Consists of a decision tree and is an extension of Random Forest in survival analysis. It predicts an individual's survival function by averaging the survival tree on multiple subsamples.

from sksurv.util import Surv
import xgboost as xgb

xgb_clf = xgb.XGBClassifier(objective='binary:logistic', eval_metric='logloss', use_label_encoder=False, random_state=42)
xgb_clf.fit(X_scaled, y['event'])

event_probabilities = xgb_clf.predict_proba(X_scaled)[:, 1]

print(event_probabilities[:5])

!conda install scipy -y

!conda update numpy scipy scikit-learn -y

!conda install -c anaconda gfortran -y
